requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
alembic==1.12.1
altair==5.4.1
annotated-types==0.7.0
apispec==6.3.0
apispec-webframeworks==1.2.0
argcomplete==3.1.6
astroid==3.0.1
asttokens==2.4.1
async-timeout==4.0.3
atlas_rucio_policy_package==0.1.1
attrs==24.2.0
bcrypt==4.2.0
black==23.12.1
blinker==1.8.2
boto3==1.29.5
botocore==1.32.7
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.1
cx-Oracle==8.3.0
cycler==0.12.1
databind==4.5.2
databind.core==4.5.2
databind.json==4.5.2
debugpy==1.8.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
docopt==0.6.2
docspec==2.2.1
docspec-python==2.2.1
docstring_parser==0.11
dogpile.cache==1.2.2
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
filelock==3.16.1
flake8==6.1.0
Flask==3.0.0
fonttools==4.54.1
frozenlist==1.4.1
future==1.0.0
geoip2==4.7.0
gitdb==4.0.11
GitPython==3.1.43
globus-sdk==3.32.0
google-auth==2.23.4
greenlet==3.1.1
gssapi==1.8.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipynb==0.5.1
ipython==8.18.1
isodate==0.6.1
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kerberos==1.3.1
kiwisolver==1.4.7
krb5==0.6.0
libtorrent==2.0.9
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
maxminddb==2.6.2
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.11.1
nest-asyncio==1.6.0
nr-date==2.1.0
nr-stream==1.1.5
nr.util==0.8.12
numpy==2.0.2
oic==1.6.1
packaging==24.1
pandas==2.2.3
paramiko==3.4.1
parso==0.8.4
pathspec==0.12.1
pbr==6.1.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus-client==0.19.0
prompt_toolkit==3.0.47
protobuf==5.28.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.11.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
pydoc-markdown==4.8.2
pyflakes==3.1.0
Pygments==2.18.0
pyjwkest==1.4.2
PyJWT==2.9.0
pykerberos==1.2.4
pylint==3.0.2
pymemcache==4.0.0
pymongo==4.6.3
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==3.1.4
pyspnego==0.11.1
pytest==7.4.3
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-magic==0.4.27
python-swiftclient==4.4.0
python3-saml==1.16.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.2.0
qbittorrent-api==2023.11.57
redis==5.0.1
referencing==0.35.1
requests==2.31.0
requests-kerberos==0.15.0
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
s3transfer==0.7.0
sh==2.0.7
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
stack-data==0.6.3
statsd==4.0.1
stevedore==5.3.0
stomp.py==8.1.0
streamlit==1.39.0
tabulate==0.9.0
tenacity==9.0.0
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
tomlkit==0.13.2
tornado==6.4.1
traitlets==5.14.3
typeapi==2.2.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
virtualenv==20.24.7
watchdog==5.0.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.4
wrapt==1.16.0
xmlsec==1.3.13
xmltodict==0.13.0
yapf==0.40.2
yarl==1.12.1
zipp==3.20.2
